# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PropFirm Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  role: UserRole;
  createdAt: string;
}

export interface AdminUser {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  role: string;
  createdAt: string;
  accountCount: number;
  totalPnl: number;
}

export interface UserStats {
  totalAccounts: number;
  activeAccounts: number;
  totalPnl: number;
  /** @nullable */
  bestAccount: string | null;
  passedAccounts: number;
}

export interface Challenge {
  id: number;
  name: string;
  accountSize: number;
  price: number;
  profitTarget: number;
  maxDailyDrawdown: number;
  maxTotalDrawdown: number;
  minTradingDays: number;
  maxTradingDays: number;
  leverage: number;
  instruments: string[];
  createdAt: string;
}

export interface ChallengeInput {
  name: string;
  accountSize: number;
  price: number;
  profitTarget: number;
  maxDailyDrawdown: number;
  maxTotalDrawdown: number;
  minTradingDays: number;
  maxTradingDays: number;
  leverage: number;
  instruments: string[];
}

export type VirtualAccountStatus =
  (typeof VirtualAccountStatus)[keyof typeof VirtualAccountStatus];

export const VirtualAccountStatus = {
  active: "active",
  passed: "passed",
  failed: "failed",
  funded: "funded",
} as const;

export interface VirtualAccount {
  id: number;
  userId: number;
  challengeId: number;
  /** @nullable */
  challengeName?: string | null;
  status: VirtualAccountStatus;
  initialBalance: number;
  currentBalance: number;
  equity: number;
  floatingPnl: number;
  totalPnl: number;
  dailyPnl: number;
  maxDrawdownReached: number;
  profitTargetReached: boolean;
  tradingDays: number;
  createdAt: string;
}

export interface EquityPoint {
  time: string;
  equity: number;
}

export interface DailyStat {
  date: string;
  pnl: number;
  trades: number;
}

export interface AccountPerformance {
  accountId: number;
  equityCurve: EquityPoint[];
  dailyStats: DailyStat[];
}

export type OrderSide = (typeof OrderSide)[keyof typeof OrderSide];

export const OrderSide = {
  buy: "buy",
  sell: "sell",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface Order {
  id: number;
  accountId: number;
  symbol: string;
  side: OrderSide;
  size: number;
  openPrice: number;
  /** @nullable */
  closePrice?: number | null;
  /** @nullable */
  currentPrice?: number | null;
  /** @nullable */
  stopLoss?: number | null;
  /** @nullable */
  takeProfit?: number | null;
  /** @nullable */
  pnl?: number | null;
  status: OrderStatus;
  openedAt: string;
  /** @nullable */
  closedAt?: string | null;
}

export type OrderInputSide =
  (typeof OrderInputSide)[keyof typeof OrderInputSide];

export const OrderInputSide = {
  buy: "buy",
  sell: "sell",
} as const;

export interface OrderInput {
  symbol: string;
  side: OrderInputSide;
  size: number;
  /** @nullable */
  stopLoss?: number | null;
  /** @nullable */
  takeProfit?: number | null;
}

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

export const PaymentStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Payment {
  id: number;
  userId: number;
  challengeId: number;
  /** @nullable */
  challengeName?: string | null;
  /** @nullable */
  userEmail?: string | null;
  amount: number;
  status: PaymentStatus;
  method: string;
  /** @nullable */
  proofUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface PaymentInput {
  challengeId: number;
  amount: number;
  method: string;
}

export interface ProofUpload {
  proofUrl: string;
}

export type PayoutRequestStatus =
  (typeof PayoutRequestStatus)[keyof typeof PayoutRequestStatus];

export const PayoutRequestStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface PayoutRequest {
  id: number;
  userId: number;
  accountId: number;
  /** @nullable */
  userEmail?: string | null;
  amount: number;
  walletAddress: string;
  status: PayoutRequestStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface PayoutInput {
  accountId: number;
  amount: number;
  walletAddress: string;
}

export interface RejectInput {
  notes: string;
}

export interface LeaderboardEntry {
  rank: number;
  displayName: string;
  accountSize: number;
  totalPnl: number;
  pnlPercent: number;
  tradingDays: number;
  status: string;
}

export interface Candle {
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface MarketPrice {
  symbol: string;
  price: number;
  change: number;
  changePercent: number;
}

export interface AdminStats {
  totalUsers: number;
  totalAccounts: number;
  activeAccounts: number;
  totalRevenue: number;
  pendingPayments: number;
  pendingPayouts: number;
  passedAccounts: number;
  failedAccounts: number;
}

export type GetMarketCandlesParams = {
  symbol: string;
  interval?: string;
};

export type GetMarketPriceParams = {
  symbol: string;
};