# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PropFirm Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /challenges/{id}:
    get:
      operationId: getChallenge
      tags: [challenges]
      summary: Get a challenge plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Challenge plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
        "404":
          description: Not found

  # ─── VIRTUAL ACCOUNTS ───────────────────────────────────────────────────────
...
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get a virtual account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Virtual account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualAccount"
        "404":
          description: Not found
  /accounts/{id}/performance:
    get:
      operationId: getAccountPerformance
      tags: [accounts]
      summary: Get account equity curve and daily performance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Performance data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountPerformance"

  # ─── VIRTUAL ORDERS ─────────────────────────────────────────────────────────
...
  /admin/payments/{id}/approve:
    post:
      operationId: adminApprovePayment
      tags: [admin]
      summary: Approve a payment and create virtual account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment approved and account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualAccount"
  /admin/payments/{id}/reject:
    post:
      operationId: adminRejectPayment
      tags: [admin]
      summary: Reject a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Payment rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
  /admin/payouts/{id}/approve:
    post:
      operationId: adminApprovePayout
      tags: [admin]
      summary: Approve a payout request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payout approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayoutRequest"
  /admin/payouts/{id}/reject:
    post:
      operationId: adminRejectPayout
      tags: [admin]
      summary: Reject a payout request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Payout rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayoutRequest"
  /admin/challenges:
    post:
      operationId: adminCreateChallenge
      tags: [admin]
      summary: Create a challenge plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeInput"
      responses:
        "201":
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
...
components:
  schemas:
...
    Challenge:
      type: object
      required: [id, name, accountSize, price, profitTarget, maxDailyDrawdown, maxTotalDrawdown, minTradingDays, maxTradingDays, leverage, instruments, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        accountSize:
          type: number
        price:
          type: number
        profitTarget:
          type: number
        maxDailyDrawdown:
          type: number
        maxTotalDrawdown:
          type: number
        minTradingDays:
          type: integer
        maxTradingDays:
          type: integer
        leverage:
          type: integer
        instruments:
          type: array
          items:
            type: string
        createdAt:
          type: string
    ChallengeInput:
      type: object
      required: [name, accountSize, price, profitTarget, maxDailyDrawdown, maxTotalDrawdown, minTradingDays, maxTradingDays, leverage, instruments]
      properties:
        name:
          type: string
        accountSize:
          type: number
        price:
          type: number
        profitTarget:
          type: number
        maxDailyDrawdown:
          type: number
        maxTotalDrawdown:
          type: number
        minTradingDays:
          type: integer
        maxTradingDays:
          type: integer
        leverage:
          type: integer
        instruments:
          type: array
          items:
            type: string
    VirtualAccount:
      type: object
      required: [id, userId, challengeId, status, initialBalance, currentBalance, equity, floatingPnl, totalPnl, dailyPnl, maxDrawdownReached, profitTargetReached, tradingDays, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        challengeId:
          type: integer
        challengeName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, passed, failed, funded]
        initialBalance:
          type: number
        currentBalance:
          type: number
        equity:
          type: number
        floatingPnl:
          type: number
        totalPnl:
          type: number
        dailyPnl:
          type: number
        maxDrawdownReached:
          type: number
        profitTargetReached:
          type: boolean
        tradingDays:
          type: integer
        createdAt:
          type: string
    AccountPerformance:
      type: object
      required: [accountId, equityCurve, dailyStats]
      properties:
        accountId:
          type: integer
        equityCurve:
          type: array
          items:
            $ref: "#/components/schemas/EquityPoint"
        dailyStats:
          type: array
          items:
            $ref: "#/components/schemas/DailyStat"
    EquityPoint:
      type: object
      required: [time, equity]
      properties:
        time:
          type: string
        equity:
          type: number
    DailyStat:
      type: object
      required: [date, pnl, trades]
      properties:
        date:
          type: string
        pnl:
          type: number
        trades:
          type: integer
...
    Payment:
      type: object
      required: [id, userId, challengeId, amount, status, method, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        challengeId:
          type: integer
        challengeName:
          type: ["string", "null"]
        userEmail:
          type: ["string", "null"]
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        method:
          type: string
        proofUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    PayoutRequest:
      type: object
      required: [id, userId, accountId, amount, walletAddress, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        accountId:
          type: integer
        userEmail:
          type: ["string", "null"]
        amount:
          type: number
        walletAddress:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    RejectInput:
      type: object
      required: [notes]
      properties:
        notes:
          type: string
...