# lib/api-zod/src/generated/types/accountPerformance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PropFirm Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DailyStat } from "./dailyStat";
import type { EquityPoint } from "./equityPoint";

export interface AccountPerformance {
  accountId: number;
  equityCurve: EquityPoint[];
  dailyStats: DailyStat[];
}