# lib/api-zod/src/generated/types/challengeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PropFirm Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ChallengeInput {
  name: string;
  accountSize: number;
  price: number;
  profitTarget: number;
  maxDailyDrawdown: number;
  maxTotalDrawdown: number;
  minTradingDays: number;
  maxTradingDays: number;
  leverage: number;
  instruments: string[];
}